Follow cluster chain while parsing directory entries
When encountering LFN entries, the cluster chain was not
followed, but instead the address has simply been incremented.
This lead to exceptions since random data was parsed as an LFN
entry.

The directory parse functions were completely refactored to
avoid this issue from happening, by always following the cluster
chain and filling a temporary "incomplete" LFN buffer that is assigned
to the next directory entry that is encountered.