Raise exception when invalid EBPB is encountered
A BS_BootSig value of 0x28 represents an earlier
version of the EBPB, where only the BS_VolId is
present, only when 0x29 is set we should check
for the BS_FilSysType. The check for FilSysType
should probably be changed to a warning in the
future.

A BS_BootSig value of 0x00 represents a pre-DOS 4.0
BPB, which does not define an EBPB at all, so the
values are zeroed out.