Do not re-order directory entries on flush (#24)
When flushing directory entries back to disk, the
files have been reordered due to a usage of `set()`
and a `walk()`-like mechanism. This had the advantage
of automatically getting rid of duplicate directory entries
on flush to disk but also had the effect that it essentially
always re-ordered files due to the nature of `set()`.

In some cases like with IO.SYS on MS-DOS, a strict ordering
is required, otherwise rendering the system unbootable.

Instead, a list of dentries is now directly passed to
the flush method, leaving ordering as-is.