Always retain last cluster when truncating a file to 0 bytes
Truncating a file to 0 bytes lead to the last
cluster in the chain being freed due to the 0
bytes requirement. Marking the last cluster in
a chain as available will not allow further modifications
to it as the cluster chain is broken and not
properly terminated with a traditional end of
cluster mark.

Instead freeing all clusters, we always keep
at least one.